requirements.txt
Django==1.11
dj-database-url==0.4.2
django-cors-headers==2.0.2
django-debug-toolbar==1.8
djangorestframework==3.6.2
gevent==1.1.2
gunicorn==19.7
psycopg2
pytz==2017.2
sqlparse==0.2.3
wheel==0.26.0
